requirements.txt
apache-airflow-providers-github
requests
apache-airflow-providers-postgres
psycopg2-binary
huggingface-hub